+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Changed `-[<LYRUIParticipantPickerDataSource> searchForParticipantsMatchingText:completion:]` to `participantPickerController:searchForParticipantsMatchingText:completion:`.
 * Changed `-[<LYRUIParticipantPickerDataSource> participants]` to `participantsForParticipantPickerController:`.
 * Changed `-[<LYRUIParticipantTableViewControllerDelegate> participantTableViewControllerDidSelectCancelButton]` to `participantTableViewControllerDidCancel:`.
+* Removed `selectionIndicator` property from `LYRUIParticipantTableViewController`.
 
 ### Bug Fixes